+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
   - *SPI change* `@Json String` database type mappers now use `@EncodedJson String` instead (#1953)
   - Sql4JSqlLogger: fix NPE when using Script
   - fix using ConstructorMapper on empty generated keys results (#1955)
+  - Handle: add new afterCommit and afterRollback transaction callbacks
 
 # 3.24.1
   - fix Bean property arguments type being erased on generic beans